2 fixes, 1) failsafe for missingblockTimer and 2) set timeSinceLastHeader every web3 switch
we want a failsafe for the retrieveBlock timer function because it seems it stops working in some cases or the provider isn't functional. we clear prev timer and restart it every web3 context switch.

We also reset the timeSinceLastHeader so it doesnt switch back right away in some cases where header hasnt responded in a long time already.